+++ BraveShared/RepresentableOptionType.swift
