+++ Sources/BraveVPN/InstallVPNView.swift
@@ -137,22 +137,3 @@ extension InstallVPNViewController {
   }
 }
 
-extension UIView {
-    /// Creates empty view with specified height or width parameter.
-    /// Used mainly to make empty space for UIStackView
-    /// Note: on iOS 11+ setCustomSpacing(value, after: View) can be used instead.
-    static func spacer(_ direction: NSLayoutConstraint.Axis, amount: Int) -> UIView {
-      let spacer = UIView()
-      spacer.snp.makeConstraints { make in
-        switch direction {
-        case .vertical:
-          make.height.equalTo(amount)
-        case .horizontal:
-          make.width.equalTo(amount)
-        @unknown default:
-          assertionFailure()
-        }
-      }
-      return spacer
-    }
-}
